+++ packages/client/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this npm module are documented in this file.
 ## [6.4.4] - 2023-08-24
 
 - Update to `monaco-vscode-api` `1.81.5`
+- Added python language client and language server (pyright) examples
 
 ## [6.4.3] - 2023-08-22